backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
geoalchemy2==0.15.2

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Redis
redis==5.2.1

# Celery
celery[redis]==5.4.0

# HTTP client (for news sources)
httpx==0.28.1
aiohttp==3.11.11

# LLM
langchain==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.1
langchain-community==0.3.14

# PDF generation
weasyprint==63.1

# Scheduling
apscheduler==3.10.4

# RSS parsing
feedparser==6.0.11

# Utilities
python-dotenv==1.0.1
